Refactored everything to support two kinds of users
