+++ lib/main.dart
@@ -1,13 +1,9 @@
 import 'package:flutter/material.dart';
-import './screens/home.dart';
-import './screens/settings.dart';
-import './screens/profile.dart';
-import 'package:main/models/user.dart';
-import 'package:main/screens/Newexplore.dart';
-import 'package:main/screens/wrapper.dart';
+import 'package:main/screens/seeker-screens/wrapper.dart';
 import 'package:main/services/auth.dart';
 import 'package:provider/provider.dart';
-import './screens/explore.dart';
+
+import 'models/user.dart';
 
 void main() => runApp(App());
