+++ lib/screens/widgets/search.dart
@@ -8,7 +8,9 @@ class MySearch extends StatelessWidget {
       body: SafeArea(
         child: Padding(
           padding: const EdgeInsets.symmetric(horizontal: 20),
-          child: SearchBar( hintText: "Try Searching Something Like Retail",),
+          child: SearchBar(
+            hintText: "Try Searching Something Like Retail",
+          ),
         ),
       ),
     );
